* JNIO networking now connects between client and server * Hash checks for network managers now work. * Fixed minor issue with the logging panel.

+++ src/taiga/code/networking/LoopbackNetwork.java
@@ -20,7 +20,6 @@
 package taiga.code.networking;
 
 import java.io.IOException;
-import java.net.DatagramPacket;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -85,7 +84,7 @@ public class LoopbackNetwork extends NetworkManager {
   }
 
   @Override
-  protected void sendPacket(Object dest, int sysid, final byte[] msg) {
+  protected void sendPacket(Object dest, short sysid, final byte[] msg) {
     exe.execute(() -> {
       if(connection != null)
         connection.packetRecieved(dest, msg, sysid);
